cli, thumbs: factor out stringifyCache
